flight model bug solved
