+++ .gitignore
@@ -6,3 +6,7 @@ __pycache__/
 poetry.lock
 *.zip
 logs/
+tensorboard_logs/
+*.txt
+.vscode/
+render_gym.py
