+++ pyproject.toml
@@ -12,6 +12,7 @@ tensorboard = "^2.4.1"
 moviepy = "^1.0.3"
 
 [tool.poetry.dev-dependencies]
+black = {version = "^20.8b1", allow-prereleases = true}
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
